requirements.txt
python-telegram-bot
mysql-connector-python
requests
python-dotenv
aio-pika
sqlalchemy
pymysql